refactor(backend): declare self-management operation registry (#821)
