+++ backend/app/features/self_management_shared/capability_catalog.py
@@ -1,11 +1,10 @@
-"""First-wave self-management capability catalog."""
+"""Registered self-management capability operations."""
 
 from __future__ import annotations
 
-from app.features.self_management_shared.actor_context import (
-    SelfManagementAction,
-    SelfManagementResource,
-    SelfManagementScope,
+from app.features.self_management_shared.actor_context import SelfManagementAction
+from app.features.self_management_shared.operation_registry import (
+    list_self_management_operation_specs,
 )
 from app.features.self_management_shared.tool_gateway import (
     SelfManagementConfirmationPolicy,
@@ -13,457 +12,19 @@ from app.features.self_management_shared.tool_gateway import (
     SelfManagementSurface,
 )
 
-_SELF_ENTRY_SURFACES = frozenset(
-    {
-        SelfManagementSurface.REST,
-        SelfManagementSurface.WEB_AGENT,
-    }
-)
-_WEB_AGENT_ONLY_SURFACES = frozenset({SelfManagementSurface.WEB_AGENT})
+ALL_SELF_MANAGEMENT_OPERATIONS = {
+    spec.operation_id: spec.build_operation()
+    for spec in list_self_management_operation_specs(first_wave_only=False)
+}
 
-SELF_AGENTS_LIST = SelfManagementOperation(
-    operation_id="self.agents.list",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.READ,
-    event_name="self_agent.list.requested",
-    tool_name="self.agents.list",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="List the current user's agents.",
-)
 
-SELF_AGENTS_GET = SelfManagementOperation(
-    operation_id="self.agents.get",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.READ,
-    event_name="self_agent.get.requested",
-    tool_name="self.agents.get",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Read one current-user agent in detail.",
-)
+def _bind_operations(*operation_ids: str) -> tuple[SelfManagementOperation, ...]:
+    return tuple(
+        ALL_SELF_MANAGEMENT_OPERATIONS[operation_id] for operation_id in operation_ids
+    )
 
-SELF_AGENTS_CHECK_HEALTH = SelfManagementOperation(
-    operation_id="self.agents.check_health",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_agent.check_health.requested",
-    tool_name="self.agents.check_health",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Run a health check for one current-user agent.",
-)
 
-SELF_AGENTS_CHECK_HEALTH_ALL = SelfManagementOperation(
-    operation_id="self.agents.check_health_all",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_agent.check_health_all.requested",
-    tool_name="self.agents.check_health_all",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Run a health check sweep for all current-user agents.",
-)
-
-SELF_AGENTS_CREATE = SelfManagementOperation(
-    operation_id="self.agents.create",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_agent.create.requested",
-    tool_name="self.agents.create",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Create one current-user agent.",
-)
-
-SELF_AGENTS_UPDATE_CONFIG = SelfManagementOperation(
-    operation_id="self.agents.update_config",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_agent.update_config.requested",
-    tool_name="self.agents.update_config",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Update one current-user agent.",
-)
-
-SELF_AGENTS_DELETE = SelfManagementOperation(
-    operation_id="self.agents.delete",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_agent.delete.requested",
-    tool_name="self.agents.delete",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Soft-delete one current-user agent.",
-)
-
-SELF_AGENTS_START_SESSIONS = SelfManagementOperation(
-    operation_id="self.agents.start_sessions",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_agent.start_sessions.requested",
-    tool_name="self.agents.start_sessions",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description=(
-        "Start one or more new conversations for the current user's agents, "
-        "send a delegated message, and hand each conversation off to the "
-        "platform-managed target session without waiting for replies. In the "
-        "built-in self-management conversation, accepted target conversations "
-        "are automatically added to durable follow-up tracking."
-    ),
-)
-
-SELF_FOLLOWUPS_GET = SelfManagementOperation(
-    operation_id="self.followups.get",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.FOLLOWUPS,
-    action=SelfManagementAction.READ,
-    event_name="self_followup.get.requested",
-    tool_name="self.followups.get",
-    first_wave_exposed=True,
-    surfaces=_WEB_AGENT_ONLY_SURFACES,
-    description=(
-        "Read the current durable follow-up tracking state for the active "
-        "built-in self-management conversation, including host-managed "
-        "auto-tracked delegated targets."
-    ),
-)
-
-SELF_FOLLOWUPS_SET_SESSIONS = SelfManagementOperation(
-    operation_id="self.followups.set_sessions",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.FOLLOWUPS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_followup.set_sessions.requested",
-    tool_name="self.followups.set_sessions",
-    confirmation_policy=SelfManagementConfirmationPolicy.NONE,
-    first_wave_exposed=True,
-    surfaces=_WEB_AGENT_ONLY_SURFACES,
-    description=(
-        "Override the current tracked target conversation ids for the active "
-        "built-in self-management conversation. Use this to narrow, extend, "
-        "replace, or stop future follow-up wakeups by passing an empty list."
-    ),
-)
-
-SELF_SESSIONS_LIST = SelfManagementOperation(
-    operation_id="self.sessions.list",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.READ,
-    event_name="self_session.list.requested",
-    tool_name="self.sessions.list",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="List the current user's sessions.",
-)
-
-SELF_SESSIONS_GET = SelfManagementOperation(
-    operation_id="self.sessions.get",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.READ,
-    event_name="self_session.get.requested",
-    tool_name="self.sessions.get",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Read one current-user session in detail.",
-)
-
-SELF_SESSIONS_GET_LATEST_MESSAGES = SelfManagementOperation(
-    operation_id="self.sessions.get_latest_messages",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.READ,
-    event_name="self_session.get_latest_messages.requested",
-    tool_name="self.sessions.get_latest_messages",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description=(
-        "Read the latest persisted text messages for one or more current-user "
-        "sessions, optionally wait within a bounded budget for new target-agent "
-        "text results, and ignore reasoning, tool-call, and interrupt lifecycle "
-        "details."
-    ),
-)
-
-SELF_SESSIONS_UPDATE = SelfManagementOperation(
-    operation_id="self.sessions.update",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_session.update.requested",
-    tool_name="self.sessions.update",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Update one current-user session.",
-)
-
-SELF_SESSIONS_ARCHIVE = SelfManagementOperation(
-    operation_id="self.sessions.archive",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_session.archive.requested",
-    tool_name="self.sessions.archive",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Archive one current-user session as a soft delete.",
-)
-
-SELF_SESSIONS_UNARCHIVE = SelfManagementOperation(
-    operation_id="self.sessions.unarchive",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_session.unarchive.requested",
-    tool_name="self.sessions.unarchive",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Restore one archived current-user session.",
-)
-
-SELF_SESSIONS_SEND_MESSAGE = SelfManagementOperation(
-    operation_id="self.sessions.send_message",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.SESSIONS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_session.send_message.requested",
-    tool_name="self.sessions.send_message",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description=(
-        "Send one delegated message to one or more current-user conversations "
-        "and hand each conversation off to the platform-managed target session "
-        "without waiting for replies. In the built-in self-management "
-        "conversation, accepted target conversations are automatically added "
-        "to durable follow-up tracking."
-    ),
-)
-
-SELF_JOBS_LIST = SelfManagementOperation(
-    operation_id="self.jobs.list",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.READ,
-    event_name="self_job.list.requested",
-    tool_name="self.jobs.list",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="List the current user's jobs.",
-)
-
-SELF_JOBS_GET = SelfManagementOperation(
-    operation_id="self.jobs.get",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.READ,
-    event_name="self_job.get.requested",
-    tool_name="self.jobs.get",
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Read one current-user job in detail.",
-)
-
-SELF_JOBS_CREATE = SelfManagementOperation(
-    operation_id="self.jobs.create",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.create.requested",
-    tool_name="self.jobs.create",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description=(
-        "Create one current-user job. For `conversation_policy`, use the exact "
-        "enum `new_each_run` or `reuse_single`."
-    ),
-)
-
-SELF_JOBS_PAUSE = SelfManagementOperation(
-    operation_id="self.jobs.pause",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.pause.requested",
-    tool_name="self.jobs.pause",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Pause one current-user job.",
-)
-
-SELF_JOBS_RESUME = SelfManagementOperation(
-    operation_id="self.jobs.resume",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.resume.requested",
-    tool_name="self.jobs.resume",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Resume one current-user job.",
-)
-
-SELF_JOBS_UPDATE_PROMPT = SelfManagementOperation(
-    operation_id="self.jobs.update_prompt",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.update_prompt.requested",
-    tool_name="self.jobs.update_prompt",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Update the prompt of one current-user job.",
-)
-
-SELF_JOBS_UPDATE = SelfManagementOperation(
-    operation_id="self.jobs.update",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.update.requested",
-    tool_name="self.jobs.update",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description=(
-        "Update one current-user job. For `conversation_policy`, use the exact "
-        "enum `new_each_run` or `reuse_single`."
-    ),
-)
-
-SELF_JOBS_UPDATE_SCHEDULE = SelfManagementOperation(
-    operation_id="self.jobs.update_schedule",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.update_schedule.requested",
-    tool_name="self.jobs.update_schedule",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Update the schedule of one current-user job.",
-)
-
-SELF_JOBS_DELETE = SelfManagementOperation(
-    operation_id="self.jobs.delete",
-    scope=SelfManagementScope.SELF,
-    resource=SelfManagementResource.JOBS,
-    action=SelfManagementAction.WRITE,
-    event_name="self_job.delete.requested",
-    tool_name="self.jobs.delete",
-    confirmation_policy=SelfManagementConfirmationPolicy.REQUIRED,
-    first_wave_exposed=True,
-    surfaces=_SELF_ENTRY_SURFACES,
-    description="Soft-delete one current-user job.",
-)
-
-ADMIN_HUB_AGENTS_LIST = SelfManagementOperation(
-    operation_id="admin.agents.list",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.READ,
-    event_name="hub_agent.list.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENTS_GET = SelfManagementOperation(
-    operation_id="admin.agents.get",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.READ,
-    event_name="hub_agent.get.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENTS_CREATE = SelfManagementOperation(
-    operation_id="admin.agents.create",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="hub_agent.create.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENTS_UPDATE = SelfManagementOperation(
-    operation_id="admin.agents.update",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="hub_agent.update.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENTS_DELETE = SelfManagementOperation(
-    operation_id="admin.agents.delete",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="hub_agent.delete.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENT_ALLOWLIST_LIST = SelfManagementOperation(
-    operation_id="admin.agents.allowlist.list",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.READ,
-    event_name="hub_agent.allowlist.list.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENT_ALLOWLIST_ADD = SelfManagementOperation(
-    operation_id="admin.agents.allowlist.add",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="hub_agent.allowlist.add.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENT_ALLOWLIST_REPLACE = SelfManagementOperation(
-    operation_id="admin.agents.allowlist.replace",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="hub_agent.allowlist.replace.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-ADMIN_HUB_AGENT_ALLOWLIST_REMOVE = SelfManagementOperation(
-    operation_id="admin.agents.allowlist.remove",
-    scope=SelfManagementScope.ADMIN,
-    resource=SelfManagementResource.AGENTS,
-    action=SelfManagementAction.WRITE,
-    event_name="hub_agent.allowlist.remove.requested",
-    surfaces=frozenset({SelfManagementSurface.REST}),
-)
-
-FIRST_WAVE_EXPOSED_OPERATIONS = (
+(
     SELF_AGENTS_LIST,
     SELF_AGENTS_GET,
     SELF_AGENTS_CHECK_HEALTH,
@@ -490,9 +51,36 @@ FIRST_WAVE_EXPOSED_OPERATIONS = (
     SELF_JOBS_UPDATE_PROMPT,
     SELF_JOBS_UPDATE_SCHEDULE,
     SELF_JOBS_DELETE,
+) = _bind_operations(
+    "self.agents.list",
+    "self.agents.get",
+    "self.agents.check_health",
+    "self.agents.check_health_all",
+    "self.agents.create",
+    "self.agents.update_config",
+    "self.agents.delete",
+    "self.agents.start_sessions",
+    "self.followups.get",
+    "self.followups.set_sessions",
+    "self.sessions.list",
+    "self.sessions.get",
+    "self.sessions.get_latest_messages",
+    "self.sessions.update",
+    "self.sessions.archive",
+    "self.sessions.unarchive",
+    "self.sessions.send_message",
+    "self.jobs.list",
+    "self.jobs.get",
+    "self.jobs.create",
+    "self.jobs.pause",
+    "self.jobs.resume",
+    "self.jobs.update",
+    "self.jobs.update_prompt",
+    "self.jobs.update_schedule",
+    "self.jobs.delete",
 )
 
-INTERNAL_ADMIN_OPERATIONS = (
+(
     ADMIN_HUB_AGENTS_LIST,
     ADMIN_HUB_AGENTS_GET,
     ADMIN_HUB_AGENTS_CREATE,
@@ -502,6 +90,27 @@ INTERNAL_ADMIN_OPERATIONS = (
     ADMIN_HUB_AGENT_ALLOWLIST_ADD,
     ADMIN_HUB_AGENT_ALLOWLIST_REPLACE,
     ADMIN_HUB_AGENT_ALLOWLIST_REMOVE,
+) = _bind_operations(
+    "admin.agents.list",
+    "admin.agents.get",
+    "admin.agents.create",
+    "admin.agents.update",
+    "admin.agents.delete",
+    "admin.agents.allowlist.list",
+    "admin.agents.allowlist.add",
+    "admin.agents.allowlist.replace",
+    "admin.agents.allowlist.remove",
+)
+
+FIRST_WAVE_EXPOSED_OPERATIONS = tuple(
+    ALL_SELF_MANAGEMENT_OPERATIONS[spec.operation_id]
+    for spec in list_self_management_operation_specs()
+)
+
+INTERNAL_ADMIN_OPERATIONS = tuple(
+    ALL_SELF_MANAGEMENT_OPERATIONS[spec.operation_id]
+    for spec in list_self_management_operation_specs(first_wave_only=False)
+    if not spec.first_wave_exposed
 )
 
 UNSUPPORTED_FIRST_WAVE_OPERATION_IDS = frozenset(
@@ -510,11 +119,6 @@ UNSUPPORTED_FIRST_WAVE_OPERATION_IDS = frozenset(
         "admin.agents.delete",
     }
 )
-
-ALL_SELF_MANAGEMENT_OPERATIONS = {
-    operation.operation_id: operation
-    for operation in (*FIRST_WAVE_EXPOSED_OPERATIONS, *INTERNAL_ADMIN_OPERATIONS)
-}
 
 
 def get_self_management_operation(operation_id: str) -> SelfManagementOperation:
